requirements.txt
langchain-community
langchain
# langchain-huggingface
langchain-chroma
langchain-ollama
langchain-huggingface
pypdf
transformers
torch
sentence-transformers
accelerate
bitsandbytes
rank-bm25
docling
markitdown[pdf]
google-genai
python-dotenv